ethos-vision-prototype/backend/requirements.txt
Flask==3.0.3
requests==2.31.0
python-dotenv==1.0.1
openai==1.14.3
waitress==3.0.0
flask-cors==4.0.0
